cambios

+++ app/backend/schemas.py
@@ -262,7 +262,7 @@ class UpdateEmployeeLaborDatum(BaseModel):
     collation: int = None
     locomotion: int = None
     extra_health_payment_type_id: Union[int, None]
-    extra_health_amount: Union[int, None]
+    extra_health_amount: Union[float, None]
     apv_payment_type_id: Union[int, None]
     apv_amount: Union[int, None]
